// projects/series-movies/src/app/models/enum/sort.enum.ts
export enum SortType {
  SORT_SEARCH_RESULTS = "SORT_SEARCH_RESULTS",
  SORT_CUSTOM = "SORT_CUTOM",
  SORT_ALPHABET = "SORT_ALPHABET",
  SORT_RATING = "SORT_RATING",
  SORT_RATING_IMDB = "SORT_RATING_IMDB",
  SORT_RATING_METASCORE = "SORT_RATING_METASCORE",
  SORT_RATING_WATCHABILITY = "SORT_RATING_WATCHABILITY",
  SORT_LAST_EDITED_DATE = "SORT_LAST_EDITED_DATE",
  SORT_CREATED_DATE = "SORT_CREATED_DATE",
  SORT_YEAR = "SORT_YEAR",
  SORT_AVAILABLE_UNTIL = "SORT_AVAILABLE_UNTIL",
  SORT_RUNTIME = "SORT_RUNTIME",
  SORT_NOTE = "SORT_NOTE",
  SORT_DISCOVERY_SOURCES = "SORT_DISCOVERY_SOURCES",
  SORT_TAGLINE = "SORT_TAGLINE",
}

export enum SortKey {
  SORT_NONE = "none",
  SORT_SEARCH_RESULTS = "search-results",
  SORT_CUSTOM = "custom",
  SORT_ALPHABET = "alphabet",
  SORT_RATING = "rating",
  SORT_RATING_IMDB = "rating-imdb",
  SORT_RATING_METASCORE = "rating-metascore",
  SORT_RATING_WATCHABILITY = "rating-watchability",
  SORT_LAST_EDITED_DATE = "last-edited-date",
  SORT_CREATED_DATE = "created-date",
  SORT_YEAR = "year",
  SORT_AVAILABLE_UNTIL = "available-until",
  SORT_RUNTIME = "runtime",
  SORT_NOTE = "note",
  SORT_DISCOVERY_SOURCES = "discovery-sources",
  SORT_TAGLINE = "tagline",
}